+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuCast.scala
@@ -24,7 +24,7 @@ import scala.collection.mutable.ArrayBuffer
 import ai.rapids.cudf.{BinaryOp, ColumnVector, ColumnView, DType, Scalar}
 import ai.rapids.cudf
 import com.nvidia.spark.rapids.RapidsPluginImplicits._
-import com.nvidia.spark.rapids.shims.v2.{RapidsErrorUtils, YearParseUtil}
+import com.nvidia.spark.rapids.shims.v2.{AnsiCheckUtil, YearParseUtil}
 
 import org.apache.spark.sql.catalyst.analysis.TypeCheckResult
 import org.apache.spark.sql.catalyst.expressions.{Cast, CastBase, Expression, NullIntolerant, TimeZoneAwareExpression}
@@ -440,7 +440,9 @@ object GpuCast extends Arm {
       case (FloatType | DoubleType, TimestampType) =>
         // Spark casting to timestamp from double assumes value is in microseconds
         if (ansiMode) {
-          RapidsErrorUtils.preprocessCastFloatToTimestamp(input)
+          // We are going through a util class because Spark 3.3.0+ throws an
+          // exception if the float value is nan or +/- inf where previously it didn't
+          AnsiCheckUtil.checkAnsiCastFloatToTimestamp(input)
         }
         withResource(Scalar.fromInt(1000000)) { microsPerSec =>
           withResource(input.nansToNulls()) { inputWithNansToNull =>
